Initial commit of MiSeqParser test cases
